backend/lambda-functions/requirements_postgresql.txt
psycopg2-binary==2.9.7
sendgrid==6.10.0
